Kext: Update USB Port Mapping
Use new USB Port Mapping kernel extension based on the system. USBPort-MBP11,2.kext for newer macOS (Catalina and Big Sur) and USBPortLegacy-MBP11,2.kext for legacy macOS (Mojave and earlier). Note that for macOS Monterey, because MacBookPro11,4 SMBIOS is required, thus USB Port Mapping for that SMBIOS as not been included.

Signed-off-by: Danang Galuh Tegar Prasetyo <connect@danang.id>